Refactor shadow users tests
The main reason for this change is to create real Identity Provider data
for the shadow users (backend and core) tests related to federated
users. This is needed in a subsequent patch where we set the domain for
federated users.

I ended up moving the shadow user tests out of the identity tests, as my
changes were conflicting with some of those tests and it provided a
simpler implementation.

Partial-Bug: #1642687
Partially-Implements: bp support-federated-attr
Depends-On: I08a8f3cb59150c8e9a2f90c5ea6b0aa197a03572
Change-Id: If8c8ad39c4c55a2d800bf4432411db59799e84e6